Fix wrong directory structure
