--- ie21a-class-championship/.gitignore
@@ -1,20 +1,0 @@
-# Eclipse
-.classpath
-.project
-.settings/
-
-# Intellij
-.idea/
-*.iml
-*.iws
-
-# Mac
-.DS_Store
-
-# Maven
-log/
-target/
-=======
-target/*
-target/
-/target
