+++ src/main/java/de/bkhennef/ie21a/cc/core/entities/Badge.java
