+++ src/main/java/de/bkhennef/ie21a/cc/core/package-info.java
