+++ src/test/java/de/bkhennef/ie21a/AppTest.java
@@ -1,0 +1,20 @@
+package de.bkhennef.ie21a;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
